+++ CHANGES.md
@@ -5,6 +5,7 @@
  * Update: REST endpoints removed (trading functionality moves to the cryptotrade project)
  * Update: dropped old / outdated / dead exchanges
  * Update: Some backend support dropped
+ * Feature: free-threaded python (3.14t) support
 
 ### 2.5.0 (2026-08-08)
  * Update: Python 3.12+ required, completed move to uv, `setup.py` removed